% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithm_fges}
\alias{algorithm_fges}
\title{Algorithm FGES}
\usage{
algorithm_fges(
  max_degree = 1000,
  time_lag = 0,
  faithfulness_assumed = FALSE,
  meek_verbose = FALSE,
  parallelized = FALSE,
  symmetric_first_step = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{max_degree}{Integer. The maximum degree of the graph (min = -1)
from different random starting permutations. The model with the most
optimal BIC score will be selected. Random after the first. Defaults to 1.}

\item{time_lag}{This creates a time-series model automatically with a certain
number of lags. Defaults to zero.}

\item{faithfulness_assumed}{TRUE if (one edge) faithfulness should be assumed}

\item{meek_verbose}{TRUE if verbose output for Meek rule applications should be printed or logged}

\item{parallelized}{TRUE if the search should be parallelized}

\item{symmetric_first_step}{TRUE if the first step step for FGES should do scoring for both X->Y and Y->X}

\item{verbose}{TRUE if verbose output should be printed or logged}
}
\description{
Compose the flags of the FGES algorithm
}
\details{
Implements the Fast Greedy Equivalence Search (FGES) algorithm.
This is an implementation of the Greedy Equivalence Search algorithm,
originally due to Chris Meek but developed significantly by Max Chickering.
FGES uses with some optimizations that allow it to scale accurately to
thousands of variables accurately for the sparse case. The reference for FGES is this:
The reference for Chickering's GES is this:
Chickering (2002) "Optimal structure identification with greedy search"
Journal of Machine Learning Research.
FGES works for the continuous case, the discrete case,
and the mixed continuous/discrete case, so long as a BIC score is available
for the type of data in question.
To speed things up, it has been assumed that variables X and Y with
zero correlation do not correspond to edges in the graph. This is a
restricted form of the heuristic speedup assumption, something GES does not assume.
This heuristic speedup assumption needs to be explicitly turned on using setHeuristicSpeedup(true).
Also, edges to be added or remove from the graph in the forward or backward phase,
respectively are cached, together with the ancillary information needed to do the
additions or removals, to reduce rescoring.
A number of other optimizations were also. See code for details.
This class is configured to respect knowledge of forbidden and required edges,
including knowledge of temporal tiers.
For more details, see: https://www.phil.cmu.edu/tetrad-javadocs/7.6.0/edu/cmu/tetrad/search/Fges.html
}
\references{
Ramsey, J., Glymour, M., Sanchez-Romero, R., & Glymour, C. (2017). A million variables and more: the fast greedy equivalence search algorithm for learning high-dimensional graphical causal models, with an application to functional magnetic resonance images. International journal of data science and analytics, 3, 121-129.
}
