% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_io}
\alias{data_io}
\title{Data I/O}
\usage{
data_io(
  dataset_path,
  data_type = "continuous",
  column_delimiter = "comma",
  output_folder_path,
  is_json_output
)
}
\arguments{
\item{dataset_path}{the dataset path to apply the model}

\item{data_type}{the type of data}

\item{column_delimiter}{how columns are defined in the dataset_path file}

\item{output_folder_path}{Output directory}

\item{is_json_output}{Write out graph as json.}
}
\description{
Compose the flags of the input and output data
}
