% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{knowledge_file_path}
\alias{knowledge_file_path}
\title{Knowledge File}
\usage{
knowledge_file_path(knowledge_file_path)
}
\arguments{
\item{knowledge_file_path}{The path of the manually created knowledge file}
}
\description{
Compose the flags of the Knowledge File
}
\details{
This is a simple function that only compose the flags of the knowledge file.
The creation of the knowledge file is not yet implemented and must be done by
hand.
See https://bd2kccd.github.io/docs/causal-cmd/
}
