% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{restarts}
\alias{restarts}
\title{Restarts
Should delete this code, it is already part of BOSS and I don't believe it applies to FGES like bootstraping does.}
\usage{
restarts(num_restarts)
}
\arguments{
\item{num_restarts}{--numStarts <integer> The number of restarts, random after the first (default 1)}
}
\description{
Automatically sets --allowInternalRandomness         Allow randomness inside algorithm
}
