% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_sem_bic}
\alias{score_sem_bic}
\title{Sem BIC Score}
\usage{
score_sem_bic(
  penalty_discount = 0,
  sem_bic_rule,
  sem_bic_structure_prior = 0,
  precompute_covariances
)
}
\arguments{
\item{penalty_discount}{Penalty discount (min = 0.0)}

\item{sem_bic_rule}{Lambda: 1 = Chickering, 2 = Nandy}

\item{sem_bic_structure_prior}{Structure Prior for SEM BIC (default 0)}

\item{precompute_covariances}{True if covariance matrix should be pre-computed
for tubular continuous data}
}
\description{
Compose the flags of the Sem BIC Score.
}
\details{
This is specifically a BIC score for the linear, Gaussian case,
where we include an additional penalty term, which is commonly used.

We call this the penalty discount. So our formulas has BIC = 2L - ck log N,
where L is the likelihood, c the penalty discount (usually greater than or equal to 1),
and N the sample size.

Since the assumption is that the data are distributed as Gaussian,
this reduces to BIC = -n log sigma - ck ln N, where sigma is the standard
deviation of the linear residual obtained by regressing a child variable
onto all of its parents in the model.
}
