% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrad.R
\name{tetrad}
\alias{tetrad}
\title{Tetrad}
\usage{
tetrad(
  tetrad_cmd_path,
  data_flags,
  knowledge_flags = list(),
  algorithm_flags,
  score_flags,
  bootstrapping_flags
)
}
\arguments{
\item{tetrad_cmd_path}{path to tetrad-cmd}

\item{data_flags}{use a function in R/score.R}

\item{knowledge_flags}{use a function in R/knowledge.R}

\item{algorithm_flags}{use a function in R/algorithm.R}

\item{score_flags}{use a function in R/score.R}

\item{bootstrapping_flags}{use a function in R/bootstrapping.R}
}
\description{
Base interface for Tetrad Cmd
}
